pyproject.toml
[project]
name = "hic-agent"
version = "0.1.0"
description = "A type-safe LLM agent framework with tools, skills, and hierarchical agents"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
markers = [
    "integration: Integration tests that require API keys and may take longer to run",
]
